+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `big` props on `IconBadge` component.
+
 ## [20.16.0] - 2018-06-06
 
 Feature: